services/dots-ocr/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
Pillow==10.0.1
pdf2image==1.16.3